requirements.txt
torchvision==0.6.0
torch==1.5.0
scikit_image==0.13.1
psutil==5.7.0
opencv_contrib_python==4.4.0.44
dgl==0.5.2
h5py==2.7.1
tqdm==4.43.0
numpy==1.13.3
Pillow==8.3.2
PyYAML==5.4.1
scikit_learn==0.24.2
skimage==0.0